requirements.txt
gradio 
langchain 
langchain-community 
langchain-chroma 
langchain-huggingface 
langchain-groq 
chromadb 
bs4 
python-dotenv
sentence-transformers
